Change the style before importing the node
IE starts loading the style right after the style node is imported to
the window document. Since this is the wrong style we'll get a 404.
This patch changes the style while it's still in the template to 
avoid this issue.
A better fix is to augment the template rebase we already do for images
to include css links and stop rebasing when adding to document resources
or preloading.